+++ backend/src/models/portfolio_factory/base.py
@@ -1,31 +1,12 @@
-from api import *
-import datetime
-from dateutil import parser
-from models.portfolio_factory.trading212_builder import Trading212PortfolioFactory
 
 
-class PortfolioFactory:
-  """Factory pattern to return a Portfolio instance depending on the Broker API as input"""
+class AbstractBase:
+  """
+  Abstract base class for portfolio factory builders.
+  """
 
-  @staticmethod
-  def build(broker_api):
-    """
-    Build a Portfolio instance using the Broker API
-
-    Args:
-      broker_api (object): The Broker API instance
-
-    Returns:
-      object: The Portfolio instance
-
-    Raises:
-      ValueError: If the Broker API is not valid
-    """
-    if isinstance(broker_api, Trading212):
-      return Trading212PortfolioFactory.build(broker_api)
-    else:
-      raise ValueError(f"Unknown broker api parameter: {broker_api}")
-
+class Base:
+  """Inheritable methods"""
   @staticmethod
   def format_timestamp_to_utc_iso(timestamp):
     """
@@ -54,5 +35,3 @@ class PortfolioFactory:
       timestamp_datetime = timestamp_datetime.astimezone(datetime.timezone.utc)
 
     return timestamp_datetime.isoformat().replace("+00:00", "Z")
-
-
